Change loop variable from short to int in ReverseWordsInString
Updated the loop variable type in ReverseWordsInString from short to int to prevent negative value issues and ensure compatibility with vector size.

+++ MyCppToolbox/clsString.h
@@ -615,7 +615,7 @@ public:
 		vector<string> vString = SplitString(s1, delim);
 		if (vString.empty()) return "";
 		string s2 = "";
-		for (short i = vString.size() - 1; i >= 0; i--)
+		for (int i = vString.size() - 1; i >= 0; i--)
 		{
 			s2 += vString[i] + delim;
 		}
